refactor(ethical-reasoning): modularize codebase into core, codemode, and mcp layers
Extract core types and logic into separate modules to improve maintainability and enable programmatic API usage. Create a new TypeScript API for direct integration and refactor the MCP server to act as an adapter over the core logic. Update tests to focus on the new API and ensure backward compatibility.
