+++ .trae/documents/plan_20260212_185018.md
@@ -1,0 +1,35 @@
+# Migration Plan: Ethical Reasoning to Code Mode
+
+I have analyzed the `ethical-reasoning` MCP server and the reference Code Mode architecture. I will transform the monolithic `index.ts` into a modular 3-layer architecture.
+
+## 1. Directory Structure Setup
+- Create `src/core/`, `src/codemode/`, and `src/mcp/` directories.
+- Configure `tsconfig.json` to support the new structure if needed (though standard include patterns usually work).
+
+## 2. Layer 1: Core (`src/core/`)
+- **`types.ts`**: Extract `EthicalRequestData`, `FrameworkType`, and new result interfaces.
+- **`logic.ts`**: Move `frameworkGuidance` and the core analysis logic here.
+  - **Refactor**: Ensure it returns typed objects (e.g., `AnalysisResult`), not MCP `CallToolResult` or JSON strings.
+  - **Statelessness**: Ensure the logic is pure where possible.
+
+## 3. Layer 2: Code Mode (`src/codemode/`)
+- **`index.ts`**: Create the `EthicalReasoningAPI` class.
+  - This class will import `Core` logic.
+  - It will provide a strongly-typed TypeScript method: `analyze(request: EthicalRequestData): Promise<AnalysisResult>`.
+  - It will handle state (history) if that feature needs to be preserved at the API instance level.
+
+## 4. Layer 3: MCP Adapter (`src/mcp/`)
+- **`tools.ts`**: Define `ETHICAL_REASONING_TOOL` constant using the existing schema.
+- **`server.ts`**: Create the MCP server instance.
+  - It will instantiate `EthicalReasoningAPI`.
+  - It will map `callTool` requests to `api.analyze()`.
+  - It will format the typed result into `CallToolResult` (JSON text).
+
+## 5. Entry Point & Cleanup (`src/index.ts`)
+- Update `src/index.ts` to:
+  - Export `EthicalReasoningAPI` for programmatic use.
+  - Conditionally run the MCP server (using `src/mcp/server.ts`) when executed as a script.
+
+## 6. Testing & Verification
+- **Unit Tests**: Refactor `index.test.ts` to test the `Core` logic and `Code Mode` API directly, rather than just the MCP layer.
+- **Integration**: Verify the MCP entry point still works as expected.
